backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
asyncpg==0.31.0
python-dotenv==1.2.1
pydantic==2.12.5
python-jose==3.5.0
bcrypt==4.1.3
python-multipart==0.0.21
